More organizing of done and upcoming changes
Condense the v1->v2 changes chart and make it reader-friendly;
add a list below the chart of upcoming changes not yet made.

Remove the Ecosystem Mapping Worksheet and Goals Worksheet appendices,
at least for now.  The former probably just needs to be an example of
an ecosystem map and a little bit of text about how to make one, and
the need for the latter will be easier to determine later on.

Restore the "Benefits of Open Source" section, but mark it in the
v1->v2 chart as being considered for removal.  Although one might
infer, from the fact that there's nothing in the commit logs about its
removal, that it was just accidentally left out of the \input list in
open-source-archetypes.ltx, actually I think we did discuss removing
it but just hadn't reached a conclusion yet.  So it's back, for now.