Improve fullrefs, and use pagerefs in some places
We were using \fullref inconsistently (especially when referring to
archetypes); independently of that, we were also using it in a way
that wasn't always useful in a printed document.  Update uses of
\fullref to do the needful in context, which in some cases means
including a \pageref too.

Commit f22cfd086d on the 'attempt-to-do-traditional-sectioning' branch
was intended to fix all of this in a different way, but I wasn't able
to get that working right, so I timed out and went with this instead.